added mustUse list
added a way to force a particular card or cards to appear in the game without hard-coding the whole configuration.  Could probably use some tweaking if we want to allow drafting at some point, but works for now.  You do have to make sure that the cards you are forcing are not in another list at this point.