messages box
Added a (fairly ugly for the moment) messages box.  In the future it would be nice to have links on the card names which bring up the full-sized image (like the ones in the tooltips) on mouseover or click.  For now it's just text.  Also, it clears at the end of your turn, may want to change that later.