Adventurer, fix for Mine
Aparently Mine used to both let you gain the card you were upgrading to AND get a copy in your hand, which is nice but way more powerful than it should be.  Also, Adventurer is a bit inelegant in how it shows what all was revealed (clutters up message box) but it'll do for now.